Implement GitHub Pages deployment via API
