+++ spec/dpl/providers/pages/git_spec.rb
@@ -1,4 +1,4 @@
-describe Dpl::Providers::Pages do
+describe Dpl::Providers::Pages::Git do
   let(:args)    { |e| %w(--github_token token) + args_from_description(e) }
   let(:user)    { JSON.dump(login: 'login', name: 'name', email: 'email') }
   let(:headers) { { 'Content-Type': 'application/json', 'X-OAuth-Scopes': ['repo'] } }
